+++ loose/lib/src/constants.dart
@@ -4,6 +4,5 @@ const authority = 'firestore.googleapis.com';
 
 const apiVersion = 'v1';
 
-
 const cloudPlatformScope = 'https://www.googleapis.com/auth/cloud-platform';
 const datastoreScope = 'https://www.googleapis.com/auth/datastore';