+++ loose_builder/lib/src/loose_entity_generator.dart
@@ -3,11 +3,14 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:build/src/builder/build_step.dart';
 import 'package:loose_builder/src/from_firestore_field_helpers.dart';
 import 'package:loose_builder/src/to_firestore_field_helpers.dart';
-import 'package:loose_builder/src/query_field_helpers.dart'; 
+import 'package:loose_builder/src/query_field_helpers.dart';
 import 'package:source_gen/source_gen.dart';
 
 import 'package:loose/annotations.dart';
 import 'package:loose_builder/src/loose_builder_exception.dart';
+import 'package:loose_builder/src/uses_identifier_helper.dart'
+    show usesIdentifier;
+import 'package:loose_builder/src/constants.dart' show documentIdFieldName;
 
 final _checkForResource = const TypeChecker.fromRuntime(Resource);
 final _checkForLooseField = const TypeChecker.fromRuntime(LooseField);
@@ -31,23 +34,49 @@ class LooseDocumentGenerator extends GeneratorForAnnotation<LooseDocument> {
       throw LooseBuilderException(buf.toString());
     }
 
-    var recase = annotation.peek('useCase')?.objectValue?.getField('none')?.toIntValue() ?? 0;
-    recase = annotation.peek('useCase')?.objectValue?.getField('camelCase')?.toIntValue() ?? recase;
-    recase = annotation.peek('useCase')?.objectValue?.getField('snakeCase')?.toIntValue() ?? recase;
-    recase = annotation.peek('useCase')?.objectValue?.getField('pascalCase')?.toIntValue() ?? recase;
-    recase = annotation.peek('useCase')?.objectValue?.getField('kebabCase')?.toIntValue() ?? recase;
+    var recase = annotation
+            .peek('useCase')
+            ?.objectValue
+            ?.getField('none')
+            ?.toIntValue() ??
+        0;
+    recase = annotation
+            .peek('useCase')
+            ?.objectValue
+            ?.getField('camelCase')
+            ?.toIntValue() ??
+        recase;
+    recase = annotation
+            .peek('useCase')
+            ?.objectValue
+            ?.getField('snakeCase')
+            ?.toIntValue() ??
+        recase;
+    recase = annotation
+            .peek('useCase')
+            ?.objectValue
+            ?.getField('pascalCase')
+            ?.toIntValue() ??
+        recase;
+    recase = annotation
+            .peek('useCase')
+            ?.objectValue
+            ?.getField('kebabCase')
+            ?.toIntValue() ??
+        recase;
 
     final allowNulls = annotation.peek('allowNulls')?.boolValue ?? false;
 
-    final useDefaultValues = annotation.peek('useDefaultValues')?.boolValue ?? false;
+    final useDefaultValues =
+        annotation.peek('useDefaultValues')?.boolValue ?? false;
 
     final readonlyNulls = annotation.peek('readonlyNulls')?.boolValue ?? false;
 
     if (allowNulls && useDefaultValues) {
-      throw LooseBuilderException('allowNull and useDefaultValues should not be used together on ${element.name}.');
+      throw LooseBuilderException(
+          'allowNull and useDefaultValues should not be used together on ${element.name}.');
     }
-    
-    
+
     var name = annotation.peek('document').peek('name').stringValue;
     var parent = annotation.peek('document').peek('parent');
     var collection = parent.peek('name').stringValue;
@@ -67,36 +96,47 @@ class LooseDocumentGenerator extends GeneratorForAnnotation<LooseDocument> {
     final documentName = '_\$${className}Document';
 
     final qFields = createQueryFields(element, recase);
-    
+
     if (qFields.isNotEmpty) {
       classBuf.writeln(qFields);
       classBuf.writeln('');
     }
 
-    classBuf.writeln('class $documentName extends DocumentShell<$className> implements Documenter<$documentName, $className, _\$${className}Fields> {');
+    classBuf.writeln(
+        'class $documentName extends DocumentShell<$className> implements Documenter<$documentName, $className, _\$${className}Fields> {');
     classBuf.writeln('');
-    classBuf.writeln("$documentName([$className entity, String name = '', String createdAt = '', String updatedAt = ''])");
+    classBuf.writeln(
+        "$documentName([$className entity, String name = '', String createdAt = '', String updatedAt = ''])");
     classBuf.writeln('  : super(entity, name, createdAt, updatedAt);');
     classBuf.writeln('');
     classBuf.writeln('@override');
-    classBuf.writeln("final location = DocumentInfo('$name', '$collection', '/${path.reversed.join('/')}');");
+    classBuf.writeln(
+        "final location = DocumentInfo('$name', '$collection', '/${path.reversed.join('/')}');");
     classBuf.writeln('');
     classBuf.writeln('@override');
-    classBuf.writeln('final _\$${className}Fields queryFields = _\$${className}Fields();');
+    classBuf.writeln(
+        'final _\$${className}Fields queryFields = _\$${className}Fields();');
     classBuf.writeln('');
     classBuf.writeln('@override');
-    classBuf.writeln('$documentName from($className entity) => $documentName(entity);');
+    classBuf.writeln(
+        '$documentName from($className entity) => $documentName(entity);');
     classBuf.writeln('');
     classBuf.writeln('@override');
-    classBuf.writeln('$documentName fromFirestore(Map<String, Object> m, String name, String createTime, String updateTime) {');
+    classBuf.writeln(
+        '$documentName fromFirestore(Map<String, Object> m, String name, String createTime, String updateTime) {');
     classBuf.write('final e = $className()');
     // fromFields
-    classBuf.writeln(convertFromFirestore(element, recase, allowNulls, readonlyNulls));
+    classBuf.writeln(
+        convertFromFirestore(element, recase, allowNulls, readonlyNulls));
 
+    if (usesIdentifier(element)) {
+      classBuf.write("..${documentIdFieldName} = name.split('/').last");
+    }
     classBuf.writeln(';');
 
     // !fromFields
-    classBuf.writeln(('return $documentName(e, name, createTime, updateTime);'));
+    classBuf
+        .writeln(('return $documentName(e, name, createTime, updateTime);'));
     classBuf.writeln('}');
     classBuf.writeln('');
     classBuf.writeln('@override');
@@ -104,7 +144,8 @@ class LooseDocumentGenerator extends GeneratorForAnnotation<LooseDocument> {
     classBuf.writeln('final e = entity;');
     classBuf.write("return {'fields': ");
     // toFields
-    final converted = convertToFirestore(element, recase, allowNulls, useDefaultValues);     
+    final converted =
+        convertToFirestore(element, recase, allowNulls, useDefaultValues);
     classBuf.writeln(converted);
     classBuf.writeln('};}');
     // !toFields