+++ loose_builder/lib/src/query_field_helpers.dart
@@ -8,7 +8,10 @@ import 'package:analyzer/dart/element/type.dart';
 import 'package:source_gen/source_gen.dart';
 
 import 'package:loose_builder/src/recase_helper.dart';
+import 'package:loose_builder/src/uses_identifier_helper.dart'
+    show usesIdentifier;
 import 'package:loose_builder/src/loose_builder_exception.dart';
+import 'package:loose_builder/src/constants.dart' show documentIdFieldName;
 import 'package:loose/annotations.dart';
 
 final _checkForLooseDocument = const TypeChecker.fromRuntime(LooseDocument);
@@ -42,6 +45,9 @@ String createQueryFields(ClassElement element, int recase,
   for (final klass in classElements) {
     for (final field in klass.fields) {
       if (field.isStatic || field.isSynthetic) {
+        continue;
+      }
+      if (usesIdentifier(element) && field.name == documentIdFieldName) {
         continue;
       }
       if (_checkForLooseMap.hasAnnotationOfExact(field.type.element) ||