+++ loose_builder/lib/src/to_firestore_field_helpers.dart
@@ -6,6 +6,9 @@ import 'package:loose/annotations.dart';
 
 import 'package:loose_builder/src/loose_builder_exception.dart';
 import 'package:loose_builder/src/recase_helper.dart';
+import 'package:loose_builder/src/uses_identifier_helper.dart'
+    show usesIdentifier;
+import 'package:loose_builder/src/constants.dart' show documentIdFieldName;
 
 final _checkForLooseDocument = const TypeChecker.fromRuntime(LooseDocument);
 final _checkForLooseMap = const TypeChecker.fromRuntime(LooseMap);
@@ -35,6 +38,9 @@ String convertToFirestore(ClassElement clazz, int recase, bool globalAllowNull,
   for (final klass in classElements) {
     for (final field in klass.fields) {
       if (field.isStatic || field.isSynthetic) {
+        continue;
+      }
+      if (usesIdentifier(clazz) && field.name == documentIdFieldName) {
         continue;
       }
       var name = field.name;